requirements.txt
tqdm
yacs
h5py
numpy
scipy
numba
smplx
gdown
PyYAML
joblib
pillow
trimesh
pyrender
progress
filterp
matplotlib
tensorflow
tensorboard
torchvision
scikit-image
scikit-video
opencv-python
llvmlite
git+https://github.com/mattloper/chumpy.git
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git